yolo_worker_old/requirements.txt
kafka-python==2.0.2
ultralytics==8.2.0
opencv-python-headless==4.10.0.84
numpy>=1.24,<2
